requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
Flask-Mail==0.9.1
pymongo==4.3.3
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
python-dotenv==1.0.0
Werkzeug==2.2.3

app.run(host='0.0.0.0', port=8000, debug=True)  # Use the correct port as specified by Render